spMeshAttachment_clone spWeightedMeshAttachment_clone   * 스파인 런타임의 triangles 타입 변경에 따른 메모리 할당 오류 수정

+++ spine-c/src/spine/Attachment.c
@@ -219,7 +219,7 @@ spMeshAttachment* spMeshAttachment_clone( spMeshAttachment* source )
   self->triangles = NULL;
   if ( source->triangles != NULL )
   {
-    self->triangles = MALLOC( int, source->trianglesCount );
+    self->triangles = MALLOC( unsigned short, source->trianglesCount );
     memcpy( self->triangles, source->triangles, ( sizeof( int ) * source->trianglesCount ) );
   }
   
@@ -297,7 +297,7 @@ spWeightedMeshAttachment* spWeightedMeshAttachment_clone( spWeightedMeshAttachme
   self->triangles = NULL;
   if ( source->triangles != NULL )
   {
-    self->triangles = MALLOC( int, source->trianglesCount );
+    self->triangles = MALLOC( unsigned short, source->trianglesCount );
     memcpy( self->triangles, source->triangles, ( sizeof( int ) * source->trianglesCount ) );
   }
